implemented Simulation
Had to make the decision of spending time making it able to have multiple different types of Organisms in an environment or make it simple. So I chose simple. Only one type of organism per environment.

made best a ReactiveProperty instead of ReactiveCollection. idk why I made it a ReactiveCollection in the first place.

Shorted names of type parameters